+++ inbm/installer/install-tc.sh
@@ -44,7 +44,7 @@ if [ "$EUID" -ne 0 ]; then
 fi
 
 # Ensure we're running a supported OS
-verified_os_list=("Ubuntu 18.04" "Ubuntu 20.04" "Ubuntu 21.10")
+verified_os_list=("Ubuntu 18.04" "Ubuntu 20.04" "Ubuntu 21.10" "Ubuntu 22.04")
 
 if [[ ${verified_os_list[@]} == *"$(lsb_release -rs)"* ]]; then
   OS_TYPE="Ubuntu-$(lsb_release -rs)"
@@ -217,6 +217,9 @@ if [ "$(lsb_release -rs)" == "20.04" ]; then
 elif [ "$(lsb_release -rs)" == "21.10" ]; then
   apt-get install -y tpm2-tools tpm2-abrmd
   systemctl enable --now tpm2-abrmd
+elif [ "$(lsb_release -rs)" == "22.04" ]; then
+  apt-get install -y tpm2-tools tpm2-abrmd
+  systemctl enable tpm2-abrmd
 else
   dpkg -i tpm2-tss*.deb tpm2-abrmd*.deb tpm2-tools*.deb
   ln -sf libtss2-tcti-tabrmd.so /usr/lib/libtss2-tcti-default.so