+++ inbm/integration-reloaded/test/aota/TC16_SUCCESS_PULL_IMAGE.sh
@@ -14,11 +14,11 @@ trap test_failed ERR
 
 echo "Starting TC16 test." | systemd-cat
 
-GOOD_XML='<?xml version="1.0" encoding="utf-8"?><manifest><type>ota</type><ota><header><type>aota</type><repo>remote</repo></header><type><aota name="sample-rpm"><cmd>pull</cmd><app>docker</app><version>0</version><containerTag>registry.hub.docker.com/library/hello-world</containerTag><dockerRegistry>None</dockerRegistry></aota></type></ota></manifest>'
+GOOD_XML='<?xml version="1.0" encoding="utf-8"?><manifest><type>ota</type><ota><header><type>aota</type><repo>remote</repo></header><type><aota name="sample-rpm"><cmd>pull</cmd><app>docker</app><version>0</version><containerTag>registry.hub.docker.com/library/nginx</containerTag><dockerRegistry>None</dockerRegistry></aota></type></ota></manifest>'
 
 test_echo TC16 Succeed Pull Public Image
 test_echo
 trigger_ota "${GOOD_XML}"
 listen_ota | grep 200
 test_echo Checking that import happened.
-(trtl -cmd=list | grep hello-world)
+(trtl -cmd=list | grep nginx)