+++ inbm/integration-reloaded/test/general/INSTALL_FRAMEWORK_slow.sh
@@ -6,7 +6,7 @@ cd /vagrant/input
 cp /scripts/succeed_rpm_cert.pem .
 
 # Remove/uninstall prereqs to ensure installer deals with them
-apt-get purge -y --allow-change-held-packages lxc-common mosquitto cryptsetup less docker-compose docker.io
+apt-get purge -y --allow-change-held-packages lxc-common mosquitto cryptsetup less docker-compose 
 
 # Simulate user calling the installer
 mkdir "install TC" # test install dir with spaces
@@ -19,30 +19,38 @@ apt-get purge -y docker-ce docker-ce-cli || true
 rm -rf /var/lib/apt/lists/*
 sudo -H ACCEPT_INTEL_LICENSE=true bash -x ./install-tc.sh
 
-echo "Will install TPM Simulator"
+
+# No TPM simulator for >20.04
+SIMULATOR=0
 if [ "$(lsb_release -rs)" == "20.04" ]; then
+  echo "Will install TPM Simulator"
   dpkg -i ../tpm2-simulator20.04-0.1332-1.deb
   sed -i -e 's#ConditionPathExistsGlob=/dev/tpm.##g' /lib/systemd/system/tpm2-abrmd.service
   sed -i -e 's#ExecStart=/usr/sbin/tpm2-abrmd#ExecStart=/usr/sbin/tpm2-abrmd --tcti=libtss2-tcti-mssim.so.0#g' /lib/systemd/system/tpm2-abrmd.service
-else
+  SIMULATOR=1
+elif [ "$(lsb_release -rs)" == "18.04" ]; then
+  echo "Will install TPM Simulator"	
   dpkg -i ../tpm2-simulator18.04-0.1332-1.deb
   sed -i -e 's#ExecStart=/usr/sbin/tpm2-abrmd#ExecStart=/usr/sbin/tpm2-abrmd --tcti=libtss2-tcti-mssim.so#g' /lib/systemd/system/tpm2-abrmd.service
+  SIMULATOR=1
 fi
 
-sed -i -e 's#After=dev-tpm0.device##g' /lib/systemd/system/tpm2-abrmd.service || true
-sed -i -e 's#Requires=dev-tpm0.device##g' /lib/systemd/system/tpm2-abrmd.service || true
+if [ "$SIMULATOR" == "1" ]; then
+  sed -i -e 's#After=dev-tpm0.device##g' /lib/systemd/system/tpm2-abrmd.service || true
+  sed -i -e 's#Requires=dev-tpm0.device##g' /lib/systemd/system/tpm2-abrmd.service || true
 
-# Reload / Restart Services
-systemctl daemon-reload
-systemctl enable --now tpm2-simulator
-systemctl restart tpm2-simulator
-systemctl restart tpm2-abrmd
+  # Reload / Restart Services
+  systemctl daemon-reload
+  systemctl enable --now tpm2-simulator
+  systemctl restart tpm2-simulator
+  systemctl restart tpm2-abrmd
 
-echo "Checking TPM simulator..."
-sleep 1
-tpm2_startup -c
-tpm2_clear
-/usr/bin/tc-get-tpm-passphrase >/dev/null
+  echo "Checking TPM simulator..."
+  sleep 1
+  tpm2_startup -c
+  tpm2_clear
+  /usr/bin/tc-get-tpm-passphrase >/dev/null
+fi
 
 date
 
@@ -55,7 +63,11 @@ cp /scripts/iotg_inb_developer.conf /etc/intel_manageability.conf
 cp /scripts/inb_fw_tool_info.conf /etc/firmware_tool_info.conf
 touch /etc/intel-manageability/public/cloudadapter-agent/iot-dispatcher.cfg
 # don't connect to Telit in Integration Reloaded
-sudo -H NO_CLOUD=x PROVISION_TPM=enable NO_OTA_CERT=1 LOCAL_MQTT_PORT=9999 bash -x /usr/bin/provision-tc
+if [ "$SIMULATOR" == "1"] ; then
+  sudo -H NO_CLOUD=x PROVISION_TPM=enable NO_OTA_CERT=1 LOCAL_MQTT_PORT=9999 bash -x /usr/bin/provision-tc
+else
+  sudo -H NO_CLOUD=x PROVISION_TPM=disable NO_OTA_CERT=1 LOCAL_MQTT_PORT=9999 bash -x /usr/bin/provision-tc
+fi
 
 cp /scripts/iotg_inb_developer.conf /etc/intel_manageability.conf
 cp /scripts/inb_fw_tool_info.conf /etc/firmware_tool_info.conf
@@ -64,6 +76,7 @@ touch /etc/intel-manageability/public/cloudadapter-agent/iot-dispatcher.cfg
 
 # we don't connect to Telit in Integration Reloaded so we don't expect
 # cloudadapter-agent to function (it primarily connects to Telit)
+sleep 5
 ps -G dispatcher-agent | grep dispatcher
 ps -G telemetry-agent | grep telemetry
 ps -G configuration-agent | grep configur
@@ -107,4 +120,6 @@ systemctl disable inbm-cloudadapter
 systemctl stop inbm-cloudadapter
 
 # Disable TPM lockout for integration testing.
-tpm2_dictionarylockout --setup-parameters --max-tries=4294967295 --clear-lockout
+if [ "$SIMULATOR" == "1"] ; then
+  tpm2_dictionarylockout --setup-parameters --max-tries=4294967295 --clear-lockout
+fi