+++ inbm/integration-reloaded/test/general/UPDATE_SYSTEM.sh
@@ -4,8 +4,6 @@ export DEBIAN_FRONTEND=noninteractive
 
 set -euxo pipefail
 
-# Speed up upgrade
-systemctl stop docker
 # disable slow kernel upgrades
 apt-mark hold linux-generic linux-image-generic linux-headers-generic
 
@@ -14,6 +12,3 @@ apt-mark hold linux-generic linux-image-generic linux-headers-generic
 apt-get update
 apt-get -y upgrade || ( apt-get -y -f install && apt-get -y upgrade )
 apt-get clean
-
-# Restart docker to make sure it's running
-systemctl restart docker