+++ inbm/integration-reloaded/run-interactive.sh
@@ -6,7 +6,7 @@ rm -rf ../output
 ../build.sh
 ln -sf ../output ./input
 
-# $1 is expected to be 18.04 or 20.04 for Ubuntu version
+# $1 is expected to be 20.04 or 22.04 for Ubuntu version
 ./setup.sh "$1"
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"