+++ inbm/integration-reloaded/scripts/select-ubuntu-18.04.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -euxo pipefail
-
-cp -f Vagrantfile-vm-box-18.04 Vagrantfile-vm-box