+++ inbm/tpm2-tools/tpm2-abrmd-after-install.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-set -euxo pipefail
-
-ldconfig
-udevadm control --reload-rules
-udevadm trigger
-systemctl enable tpm2-abrmd