+++ inbm/dispatcher-agent/dispatcher/remediationmanager/remediation_manager.py
@@ -8,6 +8,7 @@
 """
 
 import logging
+import re
 from ast import literal_eval
 from typing import Any, List, Optional, Tuple
 
@@ -71,7 +72,7 @@ class RemediationManager:
     def _remove_images(self, ids: Any) -> None:
         logger.debug("Removing Images...")
         for image_id in ids:
-            if image_id not in self.container_image_list:
+            if image_id in self.container_image_list:
                 self._remove_single_image(image_id)
         self.container_image_list[:] = []
 
@@ -125,11 +126,25 @@ class RemediationManager:
             if not self.ignore_dbs_results:
                 trtl = Trtl(PseudoShellRunner())
                 image_id = None
+
+                image_name = re.sub(r"and|[-,_]", ":", container_id)
+                err, out = trtl.list()
+                if err:
+                    logger.error("Error encountered while getting container ID")
+
+                if not image_name in str(out) or "DBS" in container_id:
+                    self._dispatcher_callbacks.broker_core.telemetry(
+                        'DBS Security issue raised on containerID: ' +
+                        str(container_id) + ' not present in list.')
+                    continue
+
+                if image_name in str(out) and not self.dbs_remove_image_on_failed_container:
+                    self.container_image_list.append(image_name)
+
                 if self.dbs_remove_image_on_failed_container:
                     image_id, image_name = self._get_image_id(trtl, container_id)
                     if image_id is None:
                         raise ValueError('Cannot read image ID')
-                    self.container_image_list.append(image_name)
 
                 (out, err, code) = trtl.stop_by_id(str(container_id))
                 if err is None: