+++ inbm/integration-reloaded/setup.sh
@@ -12,6 +12,8 @@ if [ "x$1" == "x18.04" ] ; then
   "$DIR"/scripts/select-ubuntu-18.04.sh
 elif [ "x$1" == "x20.04" ] ; then
   "$DIR"/scripts/select-ubuntu-20.04.sh
+elif [ "x$1" == "x22.04" ] ; then
+  "$DIR"/scripts/select-ubuntu-22.04.sh
 else # default
   "$DIR"/scripts/select-ubuntu-20.04.sh
 fi