+++ inbm/Changelog.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - RTC 517782 - Add Ubuntu 22.04 support for INBM TC including integration test
  - RTC 517781 - Allow installation on Debian bullseye
 
+### Removed
+ - RTC 517780 - Remove Ubuntu 18.04 support
+
 ## 3.0.12 - 2022-07-14
 
 ### Fixed