+++ inbm/Changelog.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## NEXT - ?
+
+## Added
+## Fixed
+
 ## 3.0.13 - 2022-08-07
 
 ### Added