+++ inbm/Changelog.md
@@ -3,7 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## NEXT - ?
+## 3.0.13 - 2022-08-07
+
+### Added
+ - RTC 517782 - Add Ubuntu 22.04 support for INBM TC including integration test
+ - RTC 517781 - Allow installation on Debian bullseye
+
+### Fixed
+ - RTC 517230 - Added fix for telemetry agent to receive latest values after performing config set OTA and restart of telemetry agent
+ - RTC 517028 - Set schema boundary limits for telemetry configuration values
+ - RTC 498253 - Fix duplicate of DBS remove operation in docker-compose OTA
+
 
 ## 3.0.12 - 2022-07-14